# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "City Shop Directory API",
        "description": "API documentation for the City Shop Directory mobile application",
        "contact": {
            "name": "API Support",
            "email": "admin@cityshops.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/user/favorites": {
            "get": {
                "tags": [
                    "Shop Favorites"
                ],
                "summary": "Get user's favorite shops",
                "operationId": "44ad96c8bba6c49d930983f9ac51e59a",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorites retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/shops/{shopId}/favorite": {
            "post": {
                "tags": [
                    "Shop Favorites"
                ],
                "summary": "Add shop to favorites",
                "operationId": "d0a5ffc9305ae1988c363e6b8cb3b421",
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shop added to favorites"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Shop Favorites"
                ],
                "summary": "Remove shop from favorites",
                "operationId": "b45a649fd71a2ea8ce80253bd56fa19f",
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shop removed from favorites"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/shops/{shopId}/is-favorite": {
            "get": {
                "tags": [
                    "Shop Favorites"
                ],
                "summary": "Check if shop is in user's favorites",
                "operationId": "eb508c9abd870a3e1b24736d95dddcaf",
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite status retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}